Added Custom Exception Handling, Logging, and Improved Unit Tests for Controller

+++ portfolio/portfolio/src/main/java/com/robear/portfolio/service/SkillService.java
@@ -1,5 +1,6 @@
 package com.robear.portfolio.service;
 
+import com.robear.portfolio.exception.SkillNotFoundException;
 import com.robear.portfolio.service.interfaces.ISkillService;
 import com.robear.portfolio.model.Skill;
 import com.robear.portfolio.enums.SkillType;
@@ -32,12 +33,21 @@ public class SkillService implements ISkillService {
 
     @Override
     public Skill getSkillById(Long id) {
-        return skillRepository.findById(id).orElse(null);
+        Skill skill = skillRepository.findById(id).orElse(null);
+        if (skill == null) {
+            throw new SkillNotFoundException(id);
+        }
+        return skill;
     }
 
     @Override
     public List<Skill> getSkillsOfType(SkillType type) {
-        return skillRepository.findByType(type);
+        List<Skill> skills =  skillRepository.findByType(type);
+        if (skills == null)
+        {
+            throw new SkillNotFoundException(type);
+        }
+        return skills;
 
         /* Keeping this here for myself, this is how I could use Java Stream to handle it.
         return skillRepository.findAll()
@@ -49,6 +59,10 @@ public class SkillService implements ISkillService {
 
     @Override
     public void deleteSkill(Long id) {
+        Skill skill = getSkillById(id);
+        if (skill == null) {
+            throw new SkillNotFoundException(id);
+        }
         skillRepository.deleteById(id);
     }
 }
